Add -sL swith to pass linker flags at the end
There are some cases where linker flags order do matter (for example for
linking libraries). In those cases DMD provide no mechanism for altering
the order. This patches tries to minimize this problem by adding an
option (-sL) to add flags at the *end* of the linker command line. This
is specially useful to include libraries *after* the standard library
passed with -debuglib or -defaultlib.