Add deprecation() to print deprecation messages
The deprecation() function works similar to the error() or warning()
functions but only prints messages if global.params.useDeprecated is
false. If it is, it uses the error() function to actually print the
messages and halt the program.

This patch also tries to make the messages consistent with the
behaviour. Some messages said that some feature was deprecated when it
has been completely removed in reality and should be plain errors or
vice versa.

Finally, deprecations are reported with Deprecation instead of Error
(and all the error messages are normalized and produced by a single
function).